% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.download_list.R
\name{as_tibble.download_list}
\alias{as_tibble.download_list}
\title{Convert download_list into a tibble}
\usage{
\method{as_tibble}{download_list}(x)
}
\arguments{
\item{x}{download_list}
}
\description{
Convert download_list into a tibble
}
