% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa_list.R
\name{get_taxa_list}
\alias{get_taxa_list}
\title{Get unified taxon list from a nested tibble of neotoma data}
\usage{
get_taxa_list(x)
}
\arguments{
\item{x}{nested tibble from}
}
\description{
Get unified taxon list from a nested tibble of neotoma data
}
\examples{
require("dplyr") 
data(Hordaland)
Hordaland_tibble <- as_tibble(Hordaland) 
get_taxa_list(Hordaland_tibble)
}
