% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hordaland}
\alias{Hordaland}
\title{Two pollen records from Hordaland}
\format{download_list}
\usage{
Hordaland
}
\description{
A download_list produced by \code{\link[neotoma]{get_download}} containing over two pollen datasets from Hordaland.
}
\examples{
data(Hordaland)
}
\keyword{datasets}
