% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.download_list.R
\name{as_tibble.download_list}
\alias{as_tibble.download_list}
\title{Convert download_list into a tibble}
\usage{
\method{as_tibble}{download_list}(x)
}
\arguments{
\item{x}{download_list}
}
\description{
Converts a download_list produced by \code{\link[neotoma]{get_download}} into a tibble. See \code{\link[neotoma]{get_download}} for more information on how to download data.
}
\examples{
require("dplyr") 
data(Hordaland)
as_tibble(Hordaland)
}
